Hazelcast: Reorganization based on review feedback
