+++ dd-java-agent/instrumentation/hazelcast-3.6/hazelcast-3.6.gradle
@@ -1,0 +1,27 @@
+muzzle {
+  pass {
+    group = "com.hazelcast"
+    module = "hazelcast-all"
+    versions = "[3.6,3.9)"
+    assertInverse = true
+  }
+}
+
+apply from: "$rootDir/gradle/java.gradle"
+
+apply plugin: 'org.unbroken-dome.test-sets'
+
+testSets {
+  latestDepTest {
+    dirName = 'test'
+  }
+}
+
+dependencies {
+  compileOnly group: 'com.hazelcast', name: 'hazelcast-all', version: '3.6'
+
+  // Using 3.8 to allow us to exercise the async capabilities that only exist in 3.8
+  testCompile group: 'com.hazelcast', name: 'hazelcast-all', version: '3.8'
+
+  latestDepTestCompile group: 'com.hazelcast', name: 'hazelcast-all', version: '3.8.+'
+}
