--- dd-java-agent/instrumentation/hazelcast/hazelcast-3.6/hazelcast-3.6.gradle
@@ -1,27 +1,0 @@
-muzzle {
-  pass {
-    group = "com.hazelcast"
-    module = "hazelcast-all"
-    versions = "[3.6,3.9)"
-    assertInverse = true
-  }
-}
-
-apply from: "$rootDir/gradle/java.gradle"
-
-apply plugin: 'org.unbroken-dome.test-sets'
-
-testSets {
-  latestDepTest {
-    dirName = 'test'
-  }
-}
-
-dependencies {
-  // Using 3.8 since it is the only version with async support; sync support works for 3.6+
-  compileOnly group: 'com.hazelcast', name: 'hazelcast-all', version: '3.8'
-
-  testCompile group: 'com.hazelcast', name: 'hazelcast-all', version: '3.8'
-
-  latestDepTestCompile group: 'com.hazelcast', name: 'hazelcast-all', version: '3.8.+'
-}
