--- dd-java-agent/instrumentation/hazelcast/hazelcast.gradle
@@ -1,5 +1,0 @@
-/*
- * Classes that are common to all versions of the Hibernate instrumentation.
- */
-
-apply from: "$rootDir/gradle/java.gradle"
