+++ settings.gradle
@@ -122,10 +122,9 @@ include ':dd-java-agent:instrumentation:grizzly-client-1.9'
 include ':dd-java-agent:instrumentation:grizzly-http-2.3.20'
 include ':dd-java-agent:instrumentation:grpc-1.5'
 include ':dd-java-agent:instrumentation:guava-10'
-include ':dd-java-agent:instrumentation:hazelcast'
-include ':dd-java-agent:instrumentation:hazelcast:hazelcast-3.6'
-include ':dd-java-agent:instrumentation:hazelcast:hazelcast-3.9'
-include ':dd-java-agent:instrumentation:hazelcast:hazelcast-4.0'
+include ':dd-java-agent:instrumentation:hazelcast-3.6'
+include ':dd-java-agent:instrumentation:hazelcast-3.9'
+include ':dd-java-agent:instrumentation:hazelcast-4.0'
 include ':dd-java-agent:instrumentation:hibernate'
 include ':dd-java-agent:instrumentation:hibernate:core-3.3'
 include ':dd-java-agent:instrumentation:hibernate:core-4.0'
